Reorganize the Reserve
