+++ frontend/src/components/QueryItems.vue
@@ -1,36 +1,18 @@
 <template>
-    <div class="recent-listing" id="items">
-      <div class="container">
-        <div class="row">
-          <div class="col-lg-12">
-            <div class="section-heading">
-              <h2>查詢空間</h2>
-            </div>
-          </div>
-          <div class="col-lg-12">
-            <div class="">
-              <div class="item">
-                <div class="row">
-                  <EventItem v-for="item in items" :key="item.name" :item="item" @openForm="openForm(item)"/>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+  <div>
+    <ReserveList :items="items" :openForm="openForm"/>
     <ReserveForm ref="reserveFormComponent"/>
-  </template>
+  </div>
+</template>
   
   <script>
-  import EventItem from './EventItem.vue';
+  import ReserveList from './ReserveList.vue';
   import ReserveForm from './ReserveForm.vue';
 
   export default {
     name: 'QueryItems',
     data() {
       return {
-        tapImage: require('@/assets/images/tap.png'),
         items: [
           {
             name: '001',
@@ -57,7 +39,7 @@
       };
     },
     components: {
-      EventItem,
+      ReserveList,
       ReserveForm
     },
     methods: {
