+++ frontend/src/views/ReservationPage.vue
@@ -1,18 +1,18 @@
 <template>
   <div>
     <head-page pageContent="已預約空間"></head-page>
-    <reservation-item></reservation-item>
+    <reserve-child-page></reserve-child-page>
   </div>
 </template>
 
 <script>
 import HeadPage from '@/components/HeadPage.vue';
-import ReservationItem from '@/components/ReservationItem.vue';
+import ReserveChildPage from '@/components/ReserveChildPage.vue';
 
 export default {
   components: {
     HeadPage,
-    ReservationItem
+    ReserveChildPage
   }
 }
 </script>
